refactor(server):  #22 renamed google auth service and adjusted login logic
