+++ server/src/modules/authentication/jwt/jwt.strategy.ts
@@ -1,7 +1,7 @@
 import { ExtractJwt, Strategy } from "passport-jwt";
 import { PassportStrategy } from "@nestjs/passport";
 import { Injectable } from "@nestjs/common";
-import { jwtConstants } from "../const/constantes";
+import { jwtConstants } from "../../../constants/constants";
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
